review round 2: no mixin, processor owns all creation logic, strings inline
- ProjectDurableObject extends plain DurableObject: the lifecycle mixin,
  DO_CATALOG registration, initialize/ensureStarted, and the structured-name
  zod schema are gone. Project DOs are addressed by the PLAIN project id
  (ctx.id.name reads it back). The subscription bootstrap moved into
  createProject (idempotent appends) — the one creation step the processor
  cannot do for itself.
- ProjectProcessor owns creation end to end with high-level deps
  ({ env, exports, workerHost, appConfig }) instead of per-step closures:
  the D1 projects projection (vanilla prepare on the one DB binding), the
  example secret, agents root, worker build, and event forwarding all live
  in the processor now. Split into contract.ts/implementation.ts per house
  pattern so the docs catalog stays Node-safe.
- create-requested is cross-posted to /projects in the "global" namespace —
  the deployment-wide audit surface for project lifecycle (asserted in the
  ingress test).
- Event type strings are inline at use sites (no more *_EVENT_TYPE consts),
  matching the slack contract style.
- Forwarding to the worker stays on workerHost (not itx.worker.processEvent):
  that path builds a fresh checkout per call — right for tool calls, absurd
  per event. Documented at the forwarding site.
